+++ src/dataExplorer/context/scripts.tsx
@@ -1,12 +1,12 @@
 import React, {FC, createContext, useContext, useCallback} from 'react'
-import {useDispatch} from 'react-redux'
-import * as api from 'src/client/scriptsRoutes'
-import {notify} from 'src/shared/actions/notifications'
-import {
-  scriptSaveFail,
-  scriptSaveSuccess,
-} from 'src/shared/copy/notifications/categories/scripts'
 import {PersistanceContext} from 'src/dataExplorer/context/persistance'
+import {CLOUD} from 'src/shared/constants'
+
+let postScript
+
+if (CLOUD) {
+  postScript = require('src/client/scriptsRoutes').postScript
+}
 
 interface ContextType {
   handleSave: (name: string, description: string) => void
@@ -19,36 +19,28 @@ const DEFAULT_CONTEXT = {
 export const ScriptContext = createContext<ContextType>(DEFAULT_CONTEXT)
 
 export const ScriptProvider: FC = ({children}) => {
-  const dispatch = useDispatch()
   const {query} = useContext(PersistanceContext)
 
   const handleSave = useCallback(
     async (name: string, description: string) => {
-      try {
-        if (api?.postScript) {
-          const resp = await api.postScript({
-            data: {
-              name: name,
-              description,
-              script: query,
-              language: 'flux',
-            },
-          })
+      if (postScript) {
+        const resp = await postScript({
+          data: {
+            name: name,
+            description,
+            script: query,
+            language: 'flux',
+          },
+        })
 
-          if (resp.status !== 201) {
-            throw new Error(resp.data.message)
-          }
-
-          dispatch(notify(scriptSaveSuccess(name)))
-        } else {
-          alert('You are in an unsupported environment')
+        if (resp.status !== 201) {
+          throw new Error(resp.data.message)
         }
-      } catch (error) {
-        dispatch(notify(scriptSaveFail(name)))
-        console.error({error})
+      } else {
+        alert('You are in an unsupported environment')
       }
     },
-    [dispatch, query]
+    [query]
   )
 
   return (
